tcfl.shell.run: add capability to return output
If output is requested, the command will return the output produced by
the command on the console. By default it will filter \r\n to the
current CRLF convention set for the target.

As an improvement, if BASH is set, the up() function disables vi/emacs
edition to avoid readline scrolling around, which messes all the
logging up.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>