WIP: tcfl.pos: renamed all domain -> image
Removed all mentions of boot_domain, domain, etc, and replaced with
pos_mode and image.

Rearranged the code in power_on_pre_boot_domain_setup() which becomes
power_on_pre_pos_setup() for it to be leaner and in preparation to
refactor it a wee bit. It now reacts to pos_mode set to pxe (to PXE
boot the POS) or local (to do a local boot).

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>