tcfl.target_ext_shell: fix multiple calls to file_copy_to()
When calling file_copy_to() multiple times, the code was by mistake
appending to the b64 accumulator on /tmp and thus the second and
subsequent files where not valid.

Added a function to remove multiple files (files_remove) and call it
before starting the transfer remove not only the destination file but
also the /tmp/file.b64 accumulator.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>