tcfl.tc: support expecter-based interaction in multiple threads
While we had the ability to run configure and methods that don't share
targets in parallel, that was limited as it made it harder on
evaluation when we are doing long deployment or evaluation processes
that require parallel target interaction.

To solve that, first make the expecter loop thread specific, to that avail:

- introduce the concept of thread-local-storage to each testcase
  instance (and their clones)

- make the expecter be thread specific

- introduce a thread-specific constructor (__thread_init__) that we
  can call when we enter a new thread running the testcase

with this, we can now run multiple deployments in parallel that are
performing target interaction without interferring with each other.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>